增加心跳帧以及RTC请求帧

+++ Universal_controllerV2.0/user_A021_E021.h
@@ -14,6 +14,7 @@
 #include"user_initialization.h"//初始化
 #include"user_crc8.h"//CRC校验
 #include"user_judgement.h"//判断
+#include "user_Set_Correlation.h"
 
 
 
@@ -72,7 +73,7 @@ static int E021_FrameEnd6 = 0x0A;		//E021的帧尾6
 static unsigned char E021_Check_Data[50];   //用来存放接收到的数据
 static int E021_Check_Length = 0;
 
-static unsigned long Delivery_time = 30;		//用于处理定时发送的新时间变量
+static int Delivery_time = 30;		//用于处理定时发送的新时间变量
 static unsigned long Delivery_oldtime = 0;		//用于处理定时发送的旧时间变量
